+++ src/main/java/com/prgm/aroundthetown/product/room/entity/RoomImage.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.member;
+package com.prgm.aroundthetown.product.room.entity;
 
 import com.prgm.aroundthetown.common.BaseEntity;
 import lombok.*;
@@ -6,14 +6,14 @@ import lombok.*;
 import javax.persistence.*;
 
 @Entity
-@Table(name = "review_image")
+@Table(name = "room_image")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class ReviewImage extends BaseEntity {
+public class RoomImage extends BaseEntity {
 
     @Id
-    @Column(name = "review_image_id")
+    @Column(name = "room_image_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
@@ -21,13 +21,13 @@ public class ReviewImage extends BaseEntity {
     private String IMAGE_PATH;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "review_id", referencedColumnName = "review_id", nullable = false)
-    private Review review;
+    @JoinColumn(name = "room_id", referencedColumnName = "room_id", nullable = false)
+    private Room room;
 
     @Builder
-    public ReviewImage(final String IMAGE_PATH, final Review review) {
+    public RoomImage(final String IMAGE_PATH, final Room room) {
         this.IMAGE_PATH = IMAGE_PATH;
-        this.review = review;
-        review.addImage(this);
+        this.room = room;
+        room.addImage(this);
     }
 }
