+++ src/main/java/com/prgm/aroundthetown/review/entity/ReviewImage.java
@@ -1,4 +1,4 @@
-package com.prgm.aroundthetown.product.accommodation.room;
+package com.prgm.aroundthetown.review.entity;
 
 import com.prgm.aroundthetown.common.BaseEntity;
 import lombok.*;
@@ -6,14 +6,14 @@ import lombok.*;
 import javax.persistence.*;
 
 @Entity
-@Table(name = "room_image")
+@Table(name = "review_image")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class RoomImage extends BaseEntity {
+public class ReviewImage extends BaseEntity {
 
     @Id
-    @Column(name = "room_image_id")
+    @Column(name = "review_image_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
@@ -21,13 +21,13 @@ public class RoomImage extends BaseEntity {
     private String IMAGE_PATH;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "room_id", referencedColumnName = "room_id", nullable = false)
-    private Room room;
+    @JoinColumn(name = "review_id", referencedColumnName = "review_id", nullable = false)
+    private Review review;
 
     @Builder
-    public RoomImage(final String IMAGE_PATH, final Room room) {
+    public ReviewImage(final String IMAGE_PATH, final Review review) {
         this.IMAGE_PATH = IMAGE_PATH;
-        this.room = room;
-        room.addImage(this);
+        this.review = review;
+        review.addImage(this);
     }
 }
